tests/wpt/tests/shared-storage/run-url-selection-operation.tentative.https.sub.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/shared-storage/resources/util.js"></script>
<script src="/fenced-frame/resources/utils.js"></script>

<body>
<script>
'use strict';

for (const resolve_to_config of [true, false]) {
  promise_test(async () => {
    const ancestor_key = token();
    let url0 = generateURL("/shared-storage/resources/frame0.html",
                           [ancestor_key]);
    let url1 = generateURL("/shared-storage/resources/frame1.html",
                           [ancestor_key]);

    await addModuleOnce("/shared-storage/resources/simple-module.js");

    let select_url_result0 = await sharedStorage.selectURL(
        "test-url-selection-operation", [{url: url0}, {url: url1}],
        {data: {'mockResult': 0}, resolveToConfig: resolve_to_config,
         keepAlive: true});
    assert_true(validateSelectURLResult(select_url_result0, resolve_to_config));
    attachFencedFrame(select_url_result0, 'opaque-ads');
    const result0 = await nextValueFromServer(ancestor_key);
    assert_equals(result0, "frame0_loaded");

    let select_url_result1 = await sharedStorage.selectURL(
        "test-url-selection-operation", [{url: url0}, {url: url1}],
        {data: {'mockResult': 1}, resolveToConfig: resolve_to_config,
         keepAlive: true});
    assert_true(validateSelectURLResult(select_url_result1, resolve_to_config));
    attachFencedFrame(select_url_result1, 'opaque-ads');
    const result1 = await nextValueFromServer(ancestor_key);
    assert_equals(result1, "frame1_loaded");

    let select_url_result2 = await sharedStorage.selectURL(
        "test-url-selection-operation", [{url: url0}, {url: url1}],
        {data: {'mockResult': -1}, resolveToConfig: resolve_to_config,
         keepAlive: true});
    assert_true(validateSelectURLResult(select_url_result2, resolve_to_config));
    attachFencedFrame(select_url_result2, 'opaque-ads');
    const result2 = await nextValueFromServer(ancestor_key);
    assert_equals(result2, "frame0_loaded");
  }, 'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
      await addModuleOnce("/shared-storage/resources/simple-module.js");
      await sharedStorage.selectURL(
            "test-url-selection-operation",
            [{url: "1"}, {url: "2"}, {url: "3"}, {url: "4"}, {url: "5"},
             {url: "6"}, {url: "7"}, {url: "8"}, {url: "9"}],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
      assert_equals(e.message, 'Length of the \"urls\" parameter is not valid.');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with urls array length too big, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
      await addModuleOnce("/shared-storage/resources/simple-module.js");
      await sharedStorage.selectURL(
          "test-url-selection-operation",
          [],
          {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
      assert_equals(e.message, 'Length of the \"urls\" parameter is not valid.');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with empty urls array, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
      await addModuleOnce("/shared-storage/resources/simple-module.js");
      await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                reportingMetadata: {
                    'click': "https://google.com"
                }
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'TypeError');
        assert_equals(true, e.message.startsWith(
            'Failed to execute \'selectURL\' on \'SharedStorage\': ' +
            'Failed to read the \'url\' property from ' +
            '\'SharedStorageUrlWithMetadata\':') &&
            e.message.endsWith('Failed to read the \'url\' property from ' +
            '\'SharedStorageUrlWithMetadata\': Required member is undefined.'));
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with missing url, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
      await addModuleOnce("/shared-storage/resources/simple-module.js");
      await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                url: "https://#"
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
        assert_equals(e.message,
                      'The url \"https://#\" is invalid.');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with invalid url, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
      await addModuleOnce("/shared-storage/resources/simple-module.js");
      await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                url: "/shared-storage/resources/frame0.html",
                reportingMetadata: {
                    'click': "https://#"
                }
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
        assert_equals(e.message,
                      'The metadata for the url at index 0 has an invalid ' +
                      'or non-HTTPS report_url parameter \"https://#\".');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with invalid report url ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
      await addModuleOnce("/shared-storage/resources/simple-module.js");
      await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                url: "/shared-storage/resources/frame0.html",
                reportingMetadata: {
                    'click': "http://google.com"
                }
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
        assert_equals(e.message, 'The metadata for the url at index 0 has an ' +
            'invalid or non-HTTPS report_url parameter \"http://google.com\".');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with http report url, ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));

  promise_test(async t => {
    try {
      await addModuleOnce("/shared-storage/resources/simple-module.js");
      await sharedStorage.selectURL(
            "test-url-selection-operation", [{
                url: "/shared-storage/resources/frame0.html",
                reportingMetadata: {}
            }],
            {resolveToConfig: resolve_to_config, keepAlive: true});
    } catch (e) {
      assert_equals(e.name, 'DataError');
        assert_equals(e.message, 'selectURL could not get reportingMetadata ' +
                                 'object attributes');
      return;
    }
    assert_unreached("did not reject");
  }, 'selectURL() with invalid reportingMetadata ' +
     'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));
}

</script>
</body>
